Sentinel-AI-Frontend/requirements.txt
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
keyring>=24.0.0
bcrypt>=4.0.0
pymongo>=4.0.0
dnspython>=2.0.0
python-dotenv>=1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.0
google-api-python-client>=2.0.0
cryptography>=40.0.0